# public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "soca web api v1.0",
        "description": "soca web API Description",
        "version": "1.0"
    },
    "paths": {
...
        "/api/Readers/{pg}": {
            "get": {
                "tags": [
                    "Readers"
                ],
                "summary": "Get Reader data",
                "operationId": "32eaa11f647b0831897382c395d5a2d6",
                "parameters": [
                    {
                        "name": "pg",
                        "in": "path",
                        "description": "Page number for pagination 45",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Readers"
                    }
                }
            }
        },
        "/api/Readers/": {
            "post": {
                "tags": [
                    "Readers"
                ],
                "summary": "Insert reader data",
                "operationId": "c08660c67a04c4ad82f0df9fcca161d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "R_NO": {
                                        "description": "Reader number",
                                        "type": "integer"
                                    },
                                    "R_MODEL": {
                                        "description": "Reader model",
                                        "type": "integer"
                                    },
                                    "R_NAME": {
                                        "description": "Reader name",
                                        "type": "string"
                                    },
                                    "R_INTERFACE": {
                                        "description": "Reader interface",
                                        "type": "string"
                                    },
                                    "R_TIMEOUT": {
                                        "description": "Reader timeout",
                                        "type": "integer"
                                    },
                                    "R_MAP_IDX": {
                                        "description": "Reader map index",
                                        "type": "integer"
                                    },
                                    "R_READ_POINTER3": {
                                        "description": "Reader read pointer 3",
                                        "type": "integer"
                                    },
                                    "R_READ_POINTER4": {
                                        "description": "Reader read pointer 4",
                                        "type": "integer"
                                    },
                                    "POLLING": {
                                        "description": "Polling value",
                                        "type": "integer"
                                    },
                                    "CREATETIME": {
                                        "description": "Creation time",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "create Users  "
                    },
                    "500": {
                        "description": "Failed to create reader"
                    }
                }
            }
        },
...
    },
...
}